Added split.py
Added the program split.py, which can take any large text document (for
example, the contents of an entire book) and split it up into multiple
smaller text documents. which describe the entire original document,
with order preserved.  The filenames and other parameters can be
customized in the global constants section at the top of the split.py
file.